Postrequestdto 분리

+++ src/main/java/com/example/leets7th/domain/post/Service/PostService.java
@@ -1,12 +1,13 @@
 package com.example.leets7th.domain.post.Service;
 
 
+import com.example.leets7th.domain.comment.service.CommentService;
 import com.example.leets7th.domain.post.domain.Post;
 import com.example.leets7th.domain.post.domain.PostRepository;
 import com.example.leets7th.domain.post.dto.PostRequestDto;
 import com.example.leets7th.domain.post.dto.PostResponseDto;
+
 import com.example.leets7th.domain.user.domain.User;
-import com.example.leets7th.domain.user.domain.UserRepository;
 import com.example.leets7th.domain.user.service.UserService;
 import com.example.leets7th.global.code.ErrorCode;
 import com.example.leets7th.global.error.GlobalException;
@@ -23,6 +24,7 @@ import java.util.List;
 public class PostService {
     private final PostRepository postRepository;
     private final UserService userService;
+    private final CommentService commentService;
 
 
     public PostResponseDto.ReadPost getPost(Long postId) {
@@ -37,14 +39,15 @@ public class PostService {
     }
 
     public List<PostResponseDto.ReadPostList> getPostList() {
-        List<Post> posts = postRepository.findAll();
+        List<Post> posts = postRepository.findPostListWithUser();
 
         List<PostResponseDto.ReadPostList> postDtos = new ArrayList<>();
 
         for(Post post : posts) {
             PostResponseDto.ReadPostList postDto = new PostResponseDto.ReadPostList(
+                    post.getId(),
                     post.getTitle(),
-                    post.getContent(),
+                    post.getUser().getName(),
                     post.getCreatedAt(),
                     post.getUpdatedAt()
             );
@@ -57,7 +60,7 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponseDto.CreatePost createPost(PostRequestDto.Write request,Long userId) {
+    public PostResponseDto.CreatePost createPost(PostRequestDto.Create request,Long userId) {
         Post post = Post.builder()
                 .title(request.title())
                 .content(request.content())
@@ -75,10 +78,15 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponseDto.UpdatePost updatePost(PostRequestDto.Write request,Long postId) {
+    public PostResponseDto.UpdatePost updatePost(PostRequestDto.Update request,Long postId,Long userId) {
         Post post = postRepository.findById(postId).orElseThrow(()-> new GlobalException(ErrorCode.POST_NOT_FOUND));
-        post.modifyPost(request.title(), request.content());
 
+
+        if(!userId.equals(post.getUser().getId())) {
+            throw new GlobalException(ErrorCode.POST_UPDATE_NO_PERMISSION);
+        }
+
+        post.modifyPost(request.title(), request.content());
 
 
         return new PostResponseDto.UpdatePost(post.getTitle(),
@@ -89,10 +97,14 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId,Long userId) {
         Post post = postRepository.findById(postId).orElseThrow(()-> new GlobalException(ErrorCode.POST_NOT_FOUND));
-        postRepository.delete(post);
 
+        if(!userId.equals(post.getUser().getId())){
+            throw new GlobalException(ErrorCode.POST_DELETE_NO_PERMISSION);
+        }
+        commentService.deleteAllCommentByPost(postId);
+        postRepository.delete(post);
     }
